refactor, split, and add neon for rng
